Current version of script for harvesting

+++ harvest_records.py
@@ -8,7 +8,9 @@ dateTimeObj = datetime.now()
 timestampStr = dateTimeObj.strftime("%d-%b-%Y")
 
 
-def harvest_records(path: str, short_name: str, real_name: str, create_publication_dicts):
+def harvest_records(path: str, short_name: str, real_name: str, create_publication_dicts,
+                    publisher=None, publication_place=None, default_language=None, time_interval=None, host_item_sysnumber=None,
+                    field_008_18_34=None):
     return_string = ''
     try:
         try:
@@ -18,22 +20,25 @@ def harvest_records(path: str, short_name: str, real_name: str, create_publicati
                 write_error_to_logfile.comment('Letztes geharvestetes Heft von ' + real_name + ': ' + str(last_item_harvested_in_last_session))
             out = open(path + short_name + '_' + timestampStr + '.mrc', 'wb')
             pub_nr = 0
-            publication_dicts, issues_harvested = create_publication_dicts(last_item_harvested_in_last_session)
+            publication_dicts, issues_harvested = create_publication_dicts(last_item_harvested_in_last_session, short_name, real_name, publisher, publication_place, default_language, time_interval,
+                                                                           host_item_sysnumber, field_008_18_34)
             for publication_dict in publication_dicts:
                 if create_new_record.check_publication_dict_for_completeness_and_validity(publication_dict):
                     created = create_new_record.create_new_record(out, publication_dict)
                     pub_nr += created
                 else:
+                    pub_nr = 0
                     break
         except Exception as e:
             write_error_to_logfile.write(e)
             pub_nr = 0
             issues_harvested = []
+        if not pub_nr:
             if os.path.exists(path + short_name + '_' + timestampStr + '.mrc'):
                 os.remove(path + short_name + '_' + timestampStr + '.mrc')
         write_error_to_logfile.comment('Es wurden ' + str(pub_nr) + ' neue Records für ' + real_name + ' erstellt.')
         return_string += 'Es wurden ' + str(pub_nr) + ' neue Records für ' + real_name + ' erstellt.\n'
-        if issues_harvested:
+        if issues_harvested and pub_nr:
             with open('records/' + short_name + '/' + short_name + '_logfile.json', 'w') as log_file:
                 log_dict = {"last_item_harvested": max(issues_harvested)}
                 json.dump(log_dict, log_file)
